fix: code style, refactor
