+++ src/modules/Booking/components/CreateBooking/form/blocks/PriceForm/useFullPriceControl.ts
