+++ src/modules/Booking/components/CreateBooking/steps/BookingScreen1.tsx
@@ -1,8 +1,8 @@
 import { yupResolver } from "@hookform/resolvers/yup"
 import { BookingModal } from "../modal/BookingModal/BookingModal"
 import { useForm } from "react-hook-form"
-import { ScreenSchema1 } from "../../model/types/BookingValidationSchema"
-import useBookingStore from "../../store/BookingStore"
+import { ScreenSchema1 } from "../../../model/types/BookingValidationSchema"
+import useBookingStore from "../../../store/BookingStore"
 import { ScheduleForm } from "../form/blocks/ScheduleForm/ScheduleForm"
 
 const BookingScreen1 = () => {
