+++ src/modules/Booking/components/CreateBooking/steps/BookingScreen2.tsx
@@ -1,10 +1,10 @@
 import { useForm } from "react-hook-form"
 import { BookingModal } from "../modal/BookingModal/BookingModal"
 import { yupResolver } from "@hookform/resolvers/yup"
-import { ScreenSchema2 } from "../../model/types/BookingValidationSchema"
-import useBookingStore from "../../store/BookingStore"
+import { ScreenSchema2 } from "../../../model/types/BookingValidationSchema"
+import useBookingStore from "../../../store/BookingStore"
 import { CategoryRoomsForm } from "../form/blocks/CategoryForm/CategoryRoomsForm"
-import { StepTitle } from "../typography/StepTitle/StepTitle"
+import { StepTitle } from "../../typography/StepTitle/StepTitle"
 
 const BookingScreen2 = () => {
   const bookingData = useBookingStore(state => state.bookingData)
