+++ src/modules/Clients/components/ClientsPage/TableWithClients.tsx
@@ -1,18 +1,10 @@
 import { OwnerDto } from "@/generated/owners.ts"
-import { CardWithClient } from "@/modules/Clients/components/ClientsPage/CardWithClient.tsx"
-import { CardWithPet } from "@/modules/Clients/components/ClientsPage/CardWithPet.tsx"
 import { TableComponent } from "@/shared/ui/TableComponent.tsx"
 import { Box } from "@mui/material"
-import {
-  mapperForValuePetTypeToAnLabel,
-  mapResponseToTableView,
-} from "../../utils.ts"
 
 import { ERROR_MESSAGES } from "@/shared/constants/errors.ts"
-import { useNavigate } from "react-router-dom"
 import { Pet } from "../../types.ts"
-import { useCallback } from "react"
-import { APP_ROUTES } from "@/routes/types.ts"
+import { useClientTableRows } from "../../hooks/useClientTableRows.tsx"
 
 interface IProps {
   isLoading: boolean
@@ -53,42 +45,7 @@ export const TableWithClients: React.FC<IProps> = ({
   isError,
   error,
 }) => {
-  const navigate = useNavigate()
-  const handleOpenClientPage = useCallback(
-    (id: string) => navigate(APP_ROUTES.client(id)),
-    [navigate]
-  )
-
-  const handleOpenPetPage = useCallback(
-    (clientId: string, petId: string) =>
-      navigate(APP_ROUTES.pet(clientId, petId)),
-    [navigate]
-  )
-  const rows = mapResponseToTableView(data).map(data => ({
-    client: (
-      <CardWithClient
-        fullName={data.client.fullName}
-        mainPhone={data.client.mainPhone}
-        optionalPhone={data.client.optionalPhone}
-        registrationDate={data.client.registrationDate}
-        id={data.client.id}
-        handleCardClick={() => handleOpenClientPage(data.client.id)}
-      />
-    ),
-    pets: (
-      <Box display="flex" flexDirection="row" flexWrap="wrap" gap="12px 16px">
-        {data.pets.map((pet, petIndex) => (
-          <CardWithPet
-            key={petIndex}
-            petName={pet.petName}
-            petType={mapperForValuePetTypeToAnLabel[pet.petType]}
-            breed={pet.breed}
-            onClick={() => handleOpenPetPage(data.client.id, String(pet.id))}
-          />
-        ))}
-      </Box>
-    ),
-  }))
+  const rows = useClientTableRows(data)
 
   return (
     <Box>
